unsigned (zero adding) right shift operator
in an expression byte and short gets type promoted to int implicitly, so it becomes of 32 bits (0 to 31 bit from right to left) from 8 bits (byte) and 16 bits (short) respectively. so we have to explicitly typecast it.

so take care of signed numbers especially -ve numbers because in 'int' there will be 31 bits and -ve numbers will have all 1 as leading bits.
so unsigned right shift will also not work so do AND with 0xFF for 8 bits and with 0xF for 4 bits.